Move pickable_data server into utils
