+++ proxy/common_neon/utils/__init__.py
@@ -1,2 +1,4 @@
 from .utils import *
 
+from .pickable_data_server import PickableDataServer, PickableDataServerUser, PickableDataClient
+
