+++ proxy/mempool/__init__.py
@@ -1,5 +1,4 @@
-from .pickable_data_server import PickableDataServer
-from .pickable_data_client import PickableDataClient
+from .mempool_client import MemPoolClient
 from .mempool_service import MemPoolService
 from .mem_pool import MemPool
